Implementação de Créditos Valores Não Inteiros
Implementação da possibilidade de inserir créditos de disciplinas sem necessáriamente serem números inteiros. Atualmente a função funciona apenas com dígitos númericos separados por ponto.

+++ database/migrations/2021_03_24_233842_create_disciplinas_table.php
@@ -19,12 +19,12 @@ class CreateDisciplinasTable extends Migration
             $table->string('tipo');
             $table->string('nome');
             $table->float('nota');
-            $table->integer('creditos');
+            $table->float('creditos');
             $table->integer('carga_horaria');
             $table->string('codigo')->nullable();
             $table->string('original_name')->nullable();
             $table->string('path')->nullable();
-            $table->integer('conversao')->nullable();
+            $table->float('conversao')->nullable();
             $table->foreignId('pedido_id')->constrained('pedidos')->onDelete('cascade');
         });
     }
